templates/requirements.txt
streamlit
opencv-python-headless
face-recognition
numpy
Pillow
